Major Refactoring: Switch entirely to double precision floating points for all calculations and also for network transfer

+++ src/main/java/de/hpi/msc/jschneider/fileHandling/reading/SequenceReader.java
@@ -12,7 +12,7 @@ public interface SequenceReader extends DataSource
 
     boolean isNull();
 
-    float[] read(long start, long length);
+    double[] read(long start, long length);
 
     SequenceReader subReader(long start, long length);
 }
